Extract HostConfigurationSelector
