+++ httpclient-spring-boot-autoconfigure/src/test/java/de/dev/eth0/springboot/httpclient/impl/proxy/ConfigurableProxySelectorTest.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2020. dev-eth0.de All rights reserved.
  */
 
-package de.dev.eth0.springboot.httpclient.proxy;
+package de.dev.eth0.springboot.httpclient.impl.proxy;
 
 import static org.assertj.core.api.Assertions.assertThat;
